Refactor paths in multiple API-related modules
This update refactors import paths across several JavaScript files, specifically those related to our API functions. The main change involved transitioning from using '@configs' to '@config' in the paths. This was due to a necessary restructuring of our project directory that aimed at improving our codebase's organization, readability, and maintainability. Future contributors should note this adjustment when working with import statements.
