+++ src/common/api/collections.ts
@@ -3,8 +3,8 @@ import { useNavigate } from 'react-router-dom';
 
 import { Collection } from '../types/collection';
 
-import { IApiResponse } from '@/configs/axios/interfaces/IApiResponse';
-import { apiService } from '@/configs/axios/services/api.service';
+import { IApiResponse } from '@/config/axios/interfaces/IApiResponse';
+import { apiService } from '@/config/axios/services/api.service';
 
 export const useCollectionsQuery = () => {
 	const query = useQuery<Collection[]>({
