+++ src/common/api/invocations.ts
@@ -4,8 +4,8 @@ import { useParams } from 'react-router-dom';
 import { useToast } from '../components/ui/use-toast';
 import { Invocation, InvocationResponse } from '../types/invocation';
 
-import { IApiResponse } from '@/configs/axios/interfaces/IApiResponse';
-import { apiService } from '@/configs/axios/services/api.service';
+import { IApiResponse } from '@/config/axios/interfaces/IApiResponse';
+import { apiService } from '@/config/axios/services/api.service';
 
 export function useInvocationQuery({ id }: { id?: string }) {
 	const query = useQuery<Invocation>({
