+++ src/main.tsx
@@ -4,7 +4,7 @@ import { RouterProvider } from 'react-router-dom';
 import './index.css';
 import Providers from './providers/Providers';
 
-import router from '@configs/router';
+import router from '@/config/router';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
 	<Providers>
